FRP.TimeFlies.SignalFunctions

Description

Signal functions are stateful, reactive, and time-dependent entities which respond to inputs by producing outputs. Signal functions are typed using signal vectors, exported by the SignalVectors module. A signal vector is formed from some combination of signals, events, and empty indices. Signal inputs update only at time steps, while events may have occurrences at any time, including between time steps. Event occurrences form a preorder with respect to indicated occurrence times, (the reflexivity condition is not satisfied).

To program with signal functions, primitive signal functions may be combined using the routing combinators to connect inputs and outputs, as well as the reactivity combinator switch to permit signal functions to change behavior, and the feedback combinator loop to permit signal functions to respond to their own output.

In this documentation the term given signal function(s) refers to signal functions passed as parameters to signal function combinators, while (in that context) signal function will refer to the signal function produced by a combinator.

Datatypes

data SF init svIn svOut

Signal functions: The first type parameter distinguishes between uninitialized (just specificied) and initialized (ready to respond to input and time) signal functions. The second and third type parameters are the input and output signal vectors, respectively.

data NonInitialized

Signal function not yet running

data Initialized

Signal function running or suspended

type :~> svIn svOut = SF NonInitialized svIn svOut

Infix type alias for signal functions

type :^: svLeft svRight = SVAppend svLeft svRight

Infix type alias for signal vectors

data SVEmpty

Empty signal vector type

data SVSignal a

Singleton signal signal vector type

data SVEvent a

Singleton event signal vector type

data SVAppend svLeft svRight

Combine two signal vector types

Basic signal functions

identity :: sv :~> sv

Identity signal function: reproduce the input exactly as the output.

constant :: a -> SVEmpty :~> SVSignal a

Constant signal function: Produce the signal

asap :: a -> SVEmpty :~> SVEvent a

Produce an event at the first time step after being switched in

after :: Double -> a -> SVEmpty :~> SVEvent a

Produce an event after some time

never :: SVEmpty :~> SVEvent a

Never produce an event

Lifting

pureSignalTransformer :: (a -> b) -> SVSignal a :~> SVSignal b

Apply the given function to every sample of a signal

pureEventTransformer :: (a -> b) -> SVEvent a :~> SVEvent b

Apply the given function to each occurrence of an event

Composition and Routing

(>>>) :: (svIn :~> svBetween) -> (svBetween :~> svOut) -> svIn :~> svOut

Produce a new signal function where the output of the first signal function is used as the input to the second signal function

first :: (svIn :~> svOut) -> (svIn :^: sv) :~> (svOut :^: sv)

Produce a new signal function where the left side of the input is used as input for the given signal function, and the right side of the input is combined unchanged with the output of the given signal function

second :: (svIn :~> svOut) -> (sv :^: svIn) :~> (sv :^: svOut)

Similar to first, except the right side of the signal vector is used as input for the given signal function, and the left side is combined unchanged with the output.

swap :: (sv1 :^: sv2) :~> (sv2 :^: sv1)

Produce the left side of the input as the right side of the output, and vice versâ.

copy :: sv :~> (sv :^: sv)

Produce the input as left and right sides of the output

ignore :: sv :~> SVEmpty

Discard the input entirely and produce an empty output.

cancelLeft :: (SVEmpty :^: sv) :~> sv

Discard the empty left side of the input and use the right side as output.

uncancelLeft :: sv :~> (SVEmpty :^: sv)

Use the input as the right side of the output and leave the left side empty

cancelRight :: (sv :^: SVEmpty) :~> sv

Discard the empty right side of the input and use the left side as the output.

uncancelRight :: sv :~> (sv :^: SVEmpty)

Use the input as the left side of the output and leave the right side empty

associate :: ((sv1 :^: sv2) :^: sv3) :~> (sv1 :^: (sv2 :^: sv3))

Take an input where the left two subvectors are associated and produce an output with the same subvectors, but where the right two subvectors are associated.

unassociate :: (sv1 :^: (sv2 :^: sv3)) :~> ((sv1 :^: sv2) :^: sv3)

Take an input where the right two subvectors are associated and produce an output with the same subvectors, but where the left two subvectors are associated.

Reactivity

switch :: (svIn :~> (svOut :^: SVEvent (SF NonInitialized svIn svOut))) -> svIn :~> svOut

Provide the input as input to the given signal function, producing the left side of the given signal function's output as output. Upon the first occurrence of the event on the right side of the given signal function's output, act as the signal function contained by the occurrence. Note that the if a switch event occurs in between time samples, the switch function will combine the signal deltas produced by the initialization of the new signal function and by the first time step of the new signal function.

Feedback

loop :: ((svIn :^: svLoop) :~> (svOut :^: svLoop)) -> svIn :~> svOut

Provide the input as the left input to the given signal function. Produce the left output of the given signal function as the output. Use the right output of the given signal function as the right input to the given signal function, thus providing a means for a signal function to act on its own output and encode a feedback loop. This function is decoupled for signals, meaning that a signal function will produce output that depends on the signal input at the previous time step, but not for events, meaning that feedback events will have an immediate effect. It is the programmer's responsibility to ensure that an feedback event does not result in an infinite set of resulting events.

Time dependence

time :: SVEmpty :~> SVSignal Double

With an empty input, produce the time since the signal function began running as a signal output.

delay :: Double -> (SVEvent a :^: SVEvent Double) :~> SVEvent a

Maintain a time delay for events. The initial argument is the initial time delay. Left input event occurrences are stored as state together with the time delay at occurrence time. After that delay has expired, a corresponding event will be produced on the output. Right input event occurrences update the delay time. Note that this means that later input events may be output before prior input events if the delay time is shortened in the intervening time.

class TimeIntegrate i

Class of values integrateable with respect to time

integrate :: TimeIntegrate i => SVSignal i :~> SVSignal i

Produce as output the rectangle rule integration of the input with respect to time.

Events

filter :: (a -> Maybe b) -> SVEvent a :~> SVEvent b

Filter event occurrences

Joining

union :: (SVEvent a :^: SVEvent a) :~> SVEvent a

Produce an event occurrence on the output corresponding to an event occurrence on either input

combineSignals :: ((a, b) -> c) -> (SVSignal a :^: SVSignal b) :~> SVSignal c

Combine both inputs at each point in time using the given function.

capture :: (SVSignal a :^: SVEvent b) :~> SVEvent a

Combine a signal and an event by producing an output event occurrence for each input event occurrence, but with the value of the signal at that time interval

Evaluation

svLeft :: SVRoutable r => r svLeft -> r (SVAppend svLeft svRight)

svRight :: SVRoutable r => r svRight -> r (SVAppend svLeft svRight)

svOcc :: a -> SVEventInput (SVEvent a)

Construct an event input

svSig :: a -> SVSignalUpdate (SVSignal a)

Construct a signal update

emptyHandler :: SVHandler out SVEmpty

A handler for an empty vector

eventHandler :: (a -> out) -> SVHandler out (SVEvent a)

A handler for event occurrences

signalHandler :: (a -> out) -> SVHandler out (SVSignal a)

A handler for a signal sample

combineHandlers :: SVHandler out svLeft -> SVHandler out svRight -> SVHandler out (SVAppend svLeft svRight)

Combine handlers

sample :: a -> SVSample (SVSignal a)

Construct a signal sample

sampleEvt :: SVSample (SVEvent a)

An event placeholder in a sample

sampleNothing :: SVSample SVEmpty

An empty placeholder in a sample

combineSamples :: SVSample svLeft -> SVSample svRight -> SVSample (SVAppend svLeft svRight)

Combine samples

data SVEventInput sv

An event input

data SVSignalUpdate sv

A signal update

data SFEvalT svIn svOut m a

A monad for evaluating signal functions

type SFEvalIO svIn svOut a = SFEvalT svIn svOut IO a

The most likely use case:

data SFEvalState m svIn svOut

SFEvalState

runSFEvalT :: SFEvalT svIn svOut m a -> SFEvalState m svIn svOut -> m (a, SFEvalState m svIn svOut)

Evaluate a signal function, whose current state is stored in an SFEvalState.

initSFEval

Arguments

:: SVHandler (m ()) svOut
The set of handlers for the output of the signal function

-> SVSample svIn
The initial signal inputs for the signal function

-> Double
The initial time of the signal function

-> SF NonInitialized svIn svOut
The signal function to evaluate

-> SFEvalState m svIn svOut

Initialize the evaluation of a signal function, producing an SFEvalState.

push :: Monad m => SVEventInput svIn -> SFEvalT svIn svOut m ()

Push an event occurrence into the input of the signal function

update :: Monad m => SVSignalUpdate svIn -> SFEvalT svIn svOut m ()

Update the signal function input

step

Arguments

:: Monad m

=> Double
The latest time (not a time delta)

-> SFEvalT svIn svOut m ()

Sample the signal function